fabienvauchelles do it up and add -s option to ecs-compose-down.sh

+++ ecs-compose-down.sh
@@ -1,5 +1,20 @@
 #!/bin/bash -euxE
 
+declare -A only=()
+while [[ $# -gt 0 ]] ; do
+  key="$1"
+  case "$key" in
+      -s|--service-prefix)
+      only+=([-$2]=1)
+      shift
+      shift
+      ;;
+      *)
+      break
+      ;;    
+  esac
+done
+
 source $(dirname $0)/ecs-utils.sh
 
 eval $(getServiceConfigs)
@@ -12,15 +27,20 @@ done
 for k in "${!hofa[@]}" ; do
     options=$(echo "${hofa[$k]}" | sed -e 's/|/ --service-configs /g')
     svcname=$(echo "${hofa[$k]}" | sed -e 's/|/-/g')
-    if test "${extant[$svcname]+isset}"; then
-      ecs-cli compose -p '' -f ${STATEDIR}/docker-compose.${STACK}.json service down$options
+    drop=${svcname:1}
+    svcgroup=-${drop%%-*}
+    if [ ${#only[@]} -eq 0 ] || test "${only[$svcgroup]+isset}" ; then
+        if test "${extant[$svcname]+isset}"; then
+            ecs-cli compose --cluster $STACK -p '' -f ${STATEDIR}/docker-compose.${STACK}.json service down$options
+        fi
     fi
 done
 
-for service in $(aws ecs list-services --cluster $STACK | jq -r '.serviceArns | .[]'); do
-    aws ecs delete-service --cluster $STACK --service $service
-done
-
-for olddef in $(aws ecs list-task-definitions | jq -r ' .taskDefinitionArns | .[] ') ; do
-    aws ecs deregister-task-definition --task-definition $olddef
-done
+if [ ${#only[@]} -eq 0 ] ; then
+    for service in $(aws ecs list-services --cluster $STACK | jq -r '.serviceArns | .[]'); do
+        aws ecs delete-service --cluster $STACK --service $service
+    done
+    for olddef in $(aws ecs list-task-definitions | jq -r ' .taskDefinitionArns | .[] ') ; do
+        aws ecs deregister-task-definition --task-definition $olddef
+    done
+fi
